pyproject.toml
[tool.poetry]
name = "notify"
version = "0.1.0"
description = ""
authors = ["Alex Rozum <rozumalex@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
aiohttp_session = "^2.9.0"
cryptography = "^3.1.1"
jinja2 = "^2.11.2"
aiohttp_jinja2 = "^1.2.0"
pyyaml = "^5.3.1"
asyncpg = "^0.21.0"
flake8 = "^3.8.4"
dotmap = "^1.3.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
